# docs/static/swagger/openApiConfig.yaml
openapi: 3.0.3
info:
  title: ''
  version: ''
  description: >
    This is a simple CRUD API application made with Express and documented with
    Swagger
  termsOfService: http://swagger.io/terms/
  contact:
    email: bourdakos1@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  x-github: https://github.com/bee-travels
servers:
  - url: http://localhost:3001/api
    description: Development server server (uses local data)
  - url: https://sandbox-api.example.com:8443/v1
    description: Sandbox server (uses test data)
...
paths:
...
  /business-clients:
    get:
      tags:
        - BusinessClients
      summary: Get a list of all business clients
      description: Returns a list of all business clients.
      operationId: getAllBusinessClients
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BusinessClient'
        '400':
          description: Error occurred while obtaining the business client list
          content:
            application/json:
              schema:
                type: object
    post:
      tags:
        - BusinessClients
      summary: Create a new business client
      description: Creates a new business client and links it to a business.
      operationId: createBusinessClient
      requestBody:
        description: The data to create a new business client
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBusinessClientRequest'
        required: true
      responses:
        '201':
          description: >-
            Business client created successfully. Returns the created business
            client.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessClient'
        '400':
          description: Error creating business client.
  /business-clients/{business_clients_id}:
    get:
      tags:
        - BusinessClients
      summary: Get a business client by ID
      description: Returns a business client.
      operationId: getBusinessClientById
      parameters:
        - name: business_clients_id
          in: path
          description: The ID of the business client
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessClient'
        '400':
          description: Error occurred while obtaining the business client
          content:
            application/json:
              schema:
                type: object
    patch:
      tags:
        - BusinessClients
      summary: Update a business client
      description: Updates a business client's information.
      operationId: updateBusinessClient
      parameters:
        - name: business_clients_id
          in: path
          description: The ID of the business client to update
          required: true
          schema:
            type: string
      requestBody:
        description: The data to update
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBusinessClientRequest'
        required: true
      responses:
        '200':
          description: >-
            Business client updated successfully. Returns the updated business
            client.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessClient'
        '400':
          description: Error updating business client.
        '404':
          description: Business client not found.
    delete:
      tags:
        - BusinessClients
      summary: Remove a business client
      description: Removes a business client by their ID.
      operationId: removeBusinessClient
      parameters:
        - name: business_clients_id
          in: path
          description: The ID of the business client to remove
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Business client removed successfully.
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Error removing business client.
  /business-clients/business/{business_id}:
    get:
      tags:
        - BusinessClients
      summary: Get business clients by business ID
      description: Returns a list of business clients for a specific business ID.
      operationId: getBusinessClientsByBusinessId
      parameters:
        - name: business_id
          in: path
          description: The ID of the business
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BusinessClient'
        '400':
          description: Error occurred while obtaining the business client list
          content:
            application/json:
              schema:
                type: object
...
  /lost_items:
    get:
      tags:
        - LostItems
      summary: Get all lost items with their documents and files
      description: >-
        Retrieves all lost items, including their associated documents and
        files.
      operationId: getAllLostItems
      responses:
        '200':
          description: Successful retrieval of lost items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LostItemsResponse'
        '500':
          description: Error retrieving lost items
  /lost_items/report:
    post:
      tags:
        - LostItems
      summary: Report a found item
      description: Reports a found item and adds it to the database.
      operationId: reportFoundItem
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FoundItem'
        required: true
      responses:
        '201':
          description: Found item reported successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FoundItem'
        '400':
          description: Error reporting found item
  /lost_items/delete/:lost_item_id:
    delete:
      tags:
        - LostItems
      summary: Delete a found item
      description: Deletes a found item from the database.
      operationId: deleteFoundItem
      parameters:
        - name: lost_item_id
          in: path
          description: The ID of the found item to delete
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Found item deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LostItems'
        '400':
          description: Error deleting found item
  /lost_items/update/:lost_item_id:
    patch:
      tags:
        - LostItems
      summary: Update a lost item
      description: Updates the details of a lost item in the database.
      operationId: updateLostItem
      parameters:
        - name: lost_item_id
          in: path
          description: The ID of the lost item to update
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLostItem'
        required: true
      responses:
        '200':
          description: Lost item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LostItem'
        '400':
          description: Error updating lost item
...